make t2Timestamp u64

+++ pallets/token-manager/src/lib.rs
@@ -763,7 +763,7 @@ impl<T: Config> Pallet<T> {
         }
 
         let t2_sender = H256::from(T::AccountToBytesConvert::into_bytes(from));
-        let t2_timestamp: u32 = pallet_timestamp::Pallet::<T>::get().saturated_into::<u32>();
+        let t2_timestamp: u64 = pallet_timestamp::Pallet::<T>::get().saturated_into::<u64>();
         let lower_params = concat_lower_data(
             lower_id,
             token_id.into(),
